+++ tests/many_generics.rs
@@ -1,4 +1,5 @@
 #![cfg_attr(not(RUSTC_LINT_REASONS_IS_STABLE), feature(lint_reasons))]
+#![allow(dead_code)]
 
 use core::{marker::PhantomPinned, pin::Pin};
 use pin_init::*;