+++ tests/ui/expand/many_generics.expanded.rs
@@ -1,4 +1,5 @@
 #![feature(lint_reasons)]
+#![allow(dead_code)]
 use core::{marker::PhantomPinned, pin::Pin};
 use pin_init::*;
 trait Bar<'a, const ID: usize = 0> {
@@ -12,6 +13,8 @@ where
     r: &'b mut [&'a mut T; SIZE],
     _pin: PhantomPinned,
 }
+/// Pin-projections of [`Foo`]
+#[allow(dead_code)]
 #[doc(hidden)]
 struct FooProjection<
     '__pin,
@@ -20,9 +23,9 @@ struct FooProjection<
     T: Bar<'b> + ?Sized + 'a,
     const SIZE: usize = 0,
 > {
-    _pin: ::core::pin::Pin<&'__pin mut PhantomPinned>,
     array: &'__pin mut [u8; 1024 * 1024],
     r: &'__pin mut &'b mut [&'a mut T; SIZE],
+    _pin: ::core::pin::Pin<&'__pin mut PhantomPinned>,
     ___pin_phantom_data: ::core::marker::PhantomData<&'__pin mut ()>,
 }
 impl<'a, 'b: 'a, T: Bar<'b> + ?Sized + 'a, const SIZE: usize> Foo<'a, 'b, T, SIZE>
@@ -43,15 +46,16 @@ where
     ) -> FooProjection<'__pin, 'a, 'b, T, SIZE> {
         let this = unsafe { ::core::pin::Pin::get_unchecked_mut(self) };
         FooProjection {
-            _pin: unsafe { ::core::pin::Pin::new_unchecked(&mut this._pin) },
             array: &mut this.array,
             r: &mut this.r,
+            _pin: unsafe { ::core::pin::Pin::new_unchecked(&mut this._pin) },
             ___pin_phantom_data: ::core::marker::PhantomData,
         }
     }
 }
 const _: () = {
-    struct __ThePinData<'a, 'b: 'a, T: Bar<'b> + ?Sized + 'a, const SIZE: usize>
+    #[doc(hidden)]
+    struct __ThePinData<'a, 'b: 'a, T: Bar<'b> + ?Sized + 'a, const SIZE: usize = 0>
     where
         T: Bar<'a, 1>,
     {
@@ -84,19 +88,11 @@ const _: () = {
     where
         T: Bar<'a, 1>,
     {
-        unsafe fn _pin<E>(
-            self,
-            slot: *mut PhantomPinned,
-            init: impl ::pin_init::PinInit<PhantomPinned, E>,
-        ) -> ::core::result::Result<(), E> {
-            unsafe { ::pin_init::PinInit::__pinned_init(init, slot) }
-        }
-        unsafe fn __project__pin<'__slot>(
-            self,
-            slot: &'__slot mut PhantomPinned,
-        ) -> ::core::pin::Pin<&'__slot mut PhantomPinned> {
-            ::core::pin::Pin::new_unchecked(slot)
-        }
+        /// # Safety
+        ///
+        /// - `slot` is a valid pointer to uninitialized memory.
+        /// - the caller does not touch `slot` when `Err` is returned, they are only permitted
+        ///   to deallocate.
         unsafe fn array<E>(
             self,
             slot: *mut [u8; 1024 * 1024],
@@ -104,12 +100,20 @@ const _: () = {
         ) -> ::core::result::Result<(), E> {
             unsafe { ::pin_init::Init::__init(init, slot) }
         }
+        /// # Safety
+        ///
+        /// `slot` points at the field `array` inside of `Foo`, which is pinned.
         unsafe fn __project_array<'__slot>(
             self,
             slot: &'__slot mut [u8; 1024 * 1024],
         ) -> &'__slot mut [u8; 1024 * 1024] {
             slot
         }
+        /// # Safety
+        ///
+        /// - `slot` is a valid pointer to uninitialized memory.
+        /// - the caller does not touch `slot` when `Err` is returned, they are only permitted
+        ///   to deallocate.
         unsafe fn r<E>(
             self,
             slot: *mut &'b mut [&'a mut T; SIZE],
@@ -117,11 +121,36 @@ const _: () = {
         ) -> ::core::result::Result<(), E> {
             unsafe { ::pin_init::Init::__init(init, slot) }
         }
+        /// # Safety
+        ///
+        /// `slot` points at the field `r` inside of `Foo`, which is pinned.
         unsafe fn __project_r<'__slot>(
             self,
             slot: &'__slot mut &'b mut [&'a mut T; SIZE],
         ) -> &'__slot mut &'b mut [&'a mut T; SIZE] {
             slot
+        }
+        /// # Safety
+        ///
+        /// - `slot` is a valid pointer to uninitialized memory.
+        /// - the caller does not touch `slot` when `Err` is returned, they are only permitted
+        ///   to deallocate.
+        /// - `slot` will not move until it is dropped, i.e. it will be pinned.
+        unsafe fn _pin<E>(
+            self,
+            slot: *mut PhantomPinned,
+            init: impl ::pin_init::PinInit<PhantomPinned, E>,
+        ) -> ::core::result::Result<(), E> {
+            unsafe { ::pin_init::PinInit::__pinned_init(init, slot) }
+        }
+        /// # Safety
+        ///
+        /// `slot` points at the field `_pin` inside of `Foo`, which is pinned.
+        unsafe fn __project__pin<'__slot>(
+            self,
+            slot: &'__slot mut PhantomPinned,
+        ) -> ::core::pin::Pin<&'__slot mut PhantomPinned> {
+            unsafe { ::core::pin::Pin::new_unchecked(slot) }
         }
     }
     unsafe impl<
@@ -152,7 +181,7 @@ const _: () = {
         type Datee = Foo<'a, 'b, T, SIZE>;
     }
     #[allow(dead_code)]
-    struct __Unpin<'__pin, 'a, 'b: 'a, T: Bar<'b> + ?Sized + 'a, const SIZE: usize>
+    struct __Unpin<'__pin, 'a, 'b: 'a, T: Bar<'b> + ?Sized + 'a, const SIZE: usize = 0>
     where
         T: Bar<'a, 1>,
     {