+++ tests/ui/expand/no_field_access.rs
@@ -1,0 +1,1 @@
+../compile-fail/init/no_field_access.rs